HW3 Q1
1. Suppose 2 teams A and B are playing a series of games and the rst
team to win 4 games wins the series. Suppose that team A has a 55%
chance of winning each game and that the outcome of each game is
independent. Hint: binomial distribution
(a) What is the probability that team A wins the series? Give an
exact result and conrm it via simulation.
(b) What is the expected number of games played? Give an exact
result and conrm it via simulation.
(c) What is the expected number of games played given that team A
wins the series? Give an exact result and conrm it via simulation.