HW4 Q3
3. Suppose Alice takes the bus to work when the weather is clear, and
drives her car to work when it is raining. The probability of rain is 20%.
The wait time for the bus follows an exponential distribution with an
average wait time of 5 minutes. Once the bus arrives it always takes
35 minutes to get to work. When Alice drives to work, her commute
time is normally distributed with a mean of 30 and standard deviation
of 5.
(a) What is the theoretical mean of Alices commute time?
(b) Use simulations to estimate the mean and variance of Alices commute
time.
(c) Estimate by simulation the probability that