HW4 Q4
4. Let X~N(0; 1) and Y~N(4; 2).
(a) Plot the distribution of the mixture distribution where X is selected
with probability p and Y with probability (1-p) for three dierent
values of p.
(b) For each value of p you chose, what is the mean and variance of
the resulting distribution?
(c) Let p = 0:5 and for n = 1; 2; 4; 8; 32 plot the distribution of the
average of n samples from the mixture distribution. What is the
distribution converging to?
(d) For each value of n, what is the mean and variance of your sampling
distribution (i.e. the variance of the averages)?